backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.3.2
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1